instrumentation/python/azure-functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

splunk-opentelemetry[all]==1.21.0

# add this package to ensure trace context is added to logs
opentelemetry-instrumentation-logging
